Renamed sdi-user-session-manager to helper
