+++ src/sdi-user-session-helper.h
